@@ -1,12 +1,15 @@
 {
   "name": "@esbuild/linux-x64",
-  "version": "0.18.20",
+  "version": "0.24.0",
   "description": "The Linux 64-bit binary for esbuild, a JavaScript bundler.",
-  "repository": "https://github.com/evanw/esbuild",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/evanw/esbuild.git"
+  },
   "license": "MIT",
   "preferUnplugged": true,
   "engines": {
-    "node": ">=12"
+    "node": ">=18"
   },
   "os": [
     "linux"
