@@ -1,42 +1,47 @@
 {
   "name": "esbuild",
-  "version": "0.18.20",
+  "version": "0.24.0",
   "description": "An extremely fast JavaScript and CSS bundler and minifier.",
-  "repository": "https://github.com/evanw/esbuild",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/evanw/esbuild.git"
+  },
   "scripts": {
     "postinstall": "node install.js"
   },
   "main": "lib/main.js",
   "types": "lib/main.d.ts",
   "engines": {
-    "node": ">=12"
+    "node": ">=18"
   },
   "bin": {
     "esbuild": "bin/esbuild"
   },
   "optionalDependencies": {
-    "@esbuild/android-arm": "0.18.20",
-    "@esbuild/android-arm64": "0.18.20",
-    "@esbuild/android-x64": "0.18.20",
-    "@esbuild/darwin-arm64": "0.18.20",
-    "@esbuild/darwin-x64": "0.18.20",
-    "@esbuild/freebsd-arm64": "0.18.20",
-    "@esbuild/freebsd-x64": "0.18.20",
-    "@esbuild/linux-arm": "0.18.20",
-    "@esbuild/linux-arm64": "0.18.20",
-    "@esbuild/linux-ia32": "0.18.20",
-    "@esbuild/linux-loong64": "0.18.20",
-    "@esbuild/linux-mips64el": "0.18.20",
-    "@esbuild/linux-ppc64": "0.18.20",
-    "@esbuild/linux-riscv64": "0.18.20",
-    "@esbuild/linux-s390x": "0.18.20",
-    "@esbuild/linux-x64": "0.18.20",
-    "@esbuild/netbsd-x64": "0.18.20",
-    "@esbuild/openbsd-x64": "0.18.20",
-    "@esbuild/sunos-x64": "0.18.20",
-    "@esbuild/win32-arm64": "0.18.20",
-    "@esbuild/win32-ia32": "0.18.20",
-    "@esbuild/win32-x64": "0.18.20"
+    "@esbuild/aix-ppc64": "0.24.0",
+    "@esbuild/android-arm": "0.24.0",
+    "@esbuild/android-arm64": "0.24.0",
+    "@esbuild/android-x64": "0.24.0",
+    "@esbuild/darwin-arm64": "0.24.0",
+    "@esbuild/darwin-x64": "0.24.0",
+    "@esbuild/freebsd-arm64": "0.24.0",
+    "@esbuild/freebsd-x64": "0.24.0",
+    "@esbuild/linux-arm": "0.24.0",
+    "@esbuild/linux-arm64": "0.24.0",
+    "@esbuild/linux-ia32": "0.24.0",
+    "@esbuild/linux-loong64": "0.24.0",
+    "@esbuild/linux-mips64el": "0.24.0",
+    "@esbuild/linux-ppc64": "0.24.0",
+    "@esbuild/linux-riscv64": "0.24.0",
+    "@esbuild/linux-s390x": "0.24.0",
+    "@esbuild/linux-x64": "0.24.0",
+    "@esbuild/netbsd-x64": "0.24.0",
+    "@esbuild/openbsd-arm64": "0.24.0",
+    "@esbuild/openbsd-x64": "0.24.0",
+    "@esbuild/sunos-x64": "0.24.0",
+    "@esbuild/win32-arm64": "0.24.0",
+    "@esbuild/win32-ia32": "0.24.0",
+    "@esbuild/win32-x64": "0.24.0"
   },
   "license": "MIT"
 }
