@@ -1,5 +1,7 @@
-require('dotenv').config(); // Load environment variables
-const app = require('./src/app'); // Import the main app configuration
+import dotenv from 'dotenv';
+dotenv.config(); // Load environment variables
+
+import app from './src/app.js'; // Import the main app configuration
 
 const port = process.env.PORT || 3000; // Use PORT from .env or default to 3000
 
