@@ -1,14 +1,21 @@
-const express = require('express');
-const authRoutes = require('./routes/authRoutes');
-const postRoutes = require('./routes/postRoutes');
+import express from 'express';
+import cors from 'cors';
+import authRoutes from './routes/authRoutes.js';
+import postRoutes from './routes/postRoutes.js';
 
 const app = express();
 
 // Middleware
+app.use(cors({
+  origin: ['http://localhost:5173', 'http://localhost:5174'], // Allow both frontends
+  methods: ['GET', 'POST', 'PUT', 'DELETE'],
+  credentials: true
+}));
+
 app.use(express.json());
 
 // Routes
 app.use('/auth', authRoutes);
 app.use('/posts', postRoutes);
 
-module.exports = app;
+export default app;
