@@ -1,10 +1,10 @@
-const express = require('express');
-const { getPosts, createPost } = require('../controllers/postController');
-const authenticateToken = require('../middlewares/authMiddleware');
+import express from 'express';
+import { getPosts, createPost } from '../controllers/postController.js';
+import authenticateToken from '../middlewares/authMiddleware.js';
 
 const router = express.Router();
 
 router.get('/', getPosts); // Fetch all posts
 router.post('/', authenticateToken, createPost); // Create a post (protected)
 
-module.exports = router;
+export default router;
