@@ -1,27 +1,19 @@
-import React, { useState, useEffect } from 'react';
+import React from "react";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import Layout from "./components/Layout";
+import HomePage from "./pages/Home";
+import AboutPage from "./pages/About";
 
 function App() {
-  const [posts, setPosts] = useState([]);
-
-  useEffect(() => {
-    // Fetch posts from your API (backend)
-    fetch('http://localhost:3000/posts')  // Adjust API URL as needed
-      .then((response) => response.json())
-      .then((data) => setPosts(data));
-  }, []);
-
   return (
-    <div className="App">
-      <h1>Blog Posts</h1>
-      <ul>
-        {posts.map((post) => (
-          <li key={post.id}>
-            <h2>{post.title}</h2>
-            <p>{post.content}</p>
-          </li>
-        ))}
-      </ul>
-    </div>
+    <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Layout />}>
+          <Route index element={<HomePage />} />
+          <Route path="about" element={<AboutPage />} />
+        </Route>
+      </Routes>
+    </BrowserRouter>
   );
 }
 
